refactor: Refactor stuff
Refactored book/journal form sections into their own files
